+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ##### Changed
 * **\[dom\]** Refactor the props methods that accept or return event handlers allow null event handlers.
+* 💥 **\[dom\]** Replace `DOMElement` with `ReactNode` and remove `DOMElement`.
 
 ### [v0.13](https://github.com/realityforge/react4j/tree/v0.13) (2017-11-13)
 [Full Changelog](https://github.com/realityforge/react4j/compare/v0.12...v0.13)